requirements.txt
altgraph==0.17.2
cffi==1.15.0
colorama==0.4.4
cryptography==36.0.0
future==0.18.2
pefile==2021.9.3
PGPy==0.5.4
pyasn1==0.4.8
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyOpenSSL==21.0.0
pywin32-ctypes==0.2.0
six==1.16.0
tqdm==4.62.3
